+++ service/internal/db/db.go
@@ -4,8 +4,6 @@ package db
 
 import (
 	"context"
-	"strings"
-
 	"github.com/cello-proj/cello/internal/types"
 
 	"github.com/upper/db/v4"
@@ -55,13 +53,14 @@ const (
 	TokenEntryDB   = "tokens"
 )
 
-func NewSQLClient(host, database, user, password string, options string) (SQLClient, error) {
+func NewSQLClient(host, database, user, password string, options map[string]string) (SQLClient, error) {
 	return SQLClient{
 		host:     host,
 		database: database,
 		user:     user,
 		password: password,
-		options:  optionsToMap(options),
+		options:  options,
+		//options:  optionsToMap(options),
 	}, nil
 }
 
@@ -186,19 +185,4 @@ func (d SQLClient) ListTokenEntries(ctx context.Context, project string) ([]Toke
 
 	err = sess.WithContext(ctx).Collection(TokenEntryDB).Find("project", project).OrderBy("-created_at").All(&res)
 	return res, err
-}
-
-func optionsToMap(options string) map[string]string {
-	opts := map[string]string{}
-	options = strings.TrimSpace(options)
-
-	if options != "" {
-		kvPairs := strings.Split(options, " ")
-		for _, entry := range kvPairs {
-			kv := strings.Split(entry, "=")
-			opts[kv[0]] = kv[1]
-		}
-	}
-
-	return opts
 }
